refactor: 구현한 모든 코드는 chess패키지로 이동, Pawn은 pieces패키지로 이동
