+++ src/main/java/com/woowatechcamp/chess/pieces/Color.java
@@ -1,4 +1,4 @@
-package com.woowatechcamp;
+package com.woowatechcamp.chess.pieces;
 
 import java.util.Arrays;
